refactor: move code to a new file
