+++ pkg/model/manifest.go
@@ -110,15 +110,12 @@ type ManifestDevs map[string]*Dev
 // ManifestBuild defines all the build section
 type ManifestBuild map[string]*BuildInfo
 
-// ManifestDependencies represents the map of dependencies at a manifest
-type ManifestDependencies map[string]*Dependency
-
 // NewManifest creates a new empty manifest
 func NewManifest() *Manifest {
 	return &Manifest{
 		Dev:           map[string]*Dev{},
 		Build:         map[string]*BuildInfo{},
-		Dependencies:  map[string]*Dependency{},
+		Dependencies:  ManifestDependencies{},
 		Deploy:        &DeployInfo{},
 		GlobalForward: []forward.GlobalForward{},
 	}
@@ -882,15 +879,6 @@ func (manifest *Manifest) ExpandEnvVars() error {
 	}
 
 	return nil
-}
-
-// Dependency represents a dependency object at the manifest
-type Dependency struct {
-	Repository   string      `json:"repository" yaml:"repository"`
-	ManifestPath string      `json:"manifest,omitempty" yaml:"manifest,omitempty"`
-	Branch       string      `json:"branch,omitempty" yaml:"branch,omitempty"`
-	Variables    Environment `json:"variables,omitempty" yaml:"variables,omitempty"`
-	Wait         bool        `json:"wait,omitempty" yaml:"wait,omitempty"`
 }
 
 // InferFromStack infers data from a stackfile
